+++ src/youwol/backends/assets/routers/__init__.py
@@ -1,3 +1,4 @@
 # relative
 from .files import router as router_files
 from .images import router as router_images
+from .raw import router as router_raw
